cancel_timer_2 returns_ok
